changing blackjack as module
