+++ test/table.py
@@ -1,14 +1,14 @@
-from table import Table
+from blackjack.table import Table
 
 import unittest
 
 
 class TestTable(unittest.TestCase):
 
-    def __init__(self):
-        self.user = Table()
-
     def test_add_player(self):
+        """[summary]
+
+        """
         raise NotImplementedError
